blog/pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["João Vitor"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
amqp = "5.3.1"
asgiref = "3.8.1"
billiard = "4.2.1"
celery = "5.5.3"
certifi = "2025.8.3"
charset-normalizer = "3.4.3"
click = "8.2.1"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1.2"
click-repl = "0.3.0"
cloudinary = "1.44.1"
cron-descriptor = "2.0.6"
dj-database-url = "3.0.1"
django = "5.2"
django-celery-beat = "2.8.1"
django-cloudinary-storage = "0.3.0"
django-timezone-field = "7.1"
django-widget-tweaks = "1.5.0"
gunicorn = "23.0.0"
idna = "3.10"
kombu = "5.5.4"
packaging = "25.0"
pillow = "11.2.1"
prompt-toolkit = "3.0.52"
psycopg2-binary = "^2.9"
python-crontab = "3.3.0"
python-dateutil = "2.9.0.post0"
python-decouple = "3.8"
python-dotenv = "1.1.1"
python-magic = "0.4.27"
redis = "5.2.1"
requests = "2.32.5"
six = "1.17.0"
sqlparse = "0.5.3"
typing-extensions = "4.15.0"
tzdata = "2025.2"
urllib3 = "2.5.0"
vine = "5.1.0"
wcwidth = "0.2.13"
whitenoise = "6.11.0"
djangorestframework = "^3.16.1"
django-cors-headers = "^4.9.0"
djangorestframework-simplejwt = "^5.5.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2,<10.0.0"
black = ">=25.12.0,<26.0.0"
isort = ">=7.0.0,<8.0.0"
flake8 = ">=7.3.0,<8.0.0"

